+++ test/mocks/esp_ota_ops.h
@@ -26,12 +26,30 @@ typedef enum {
     ESP_OTA_IMG_ABORTED,
 } esp_ota_img_states_t;
 
+inline esp_partition_t esp_ota_mock_running_partition{0};
+inline bool esp_ota_mock_has_running_partition = false;
+inline esp_ota_img_states_t esp_ota_mock_running_state = ESP_OTA_IMG_UNDEFINED;
+inline bool esp_ota_mock_mark_valid_called = false;
+inline bool esp_ota_mock_mark_invalid_called = false;
+
+inline void esp_ota_mock_reset() {
+    esp_ota_mock_has_running_partition = false;
+    esp_ota_mock_running_state = ESP_OTA_IMG_UNDEFINED;
+    esp_ota_mock_mark_valid_called = false;
+    esp_ota_mock_mark_invalid_called = false;
+}
+
+inline void esp_ota_mock_set_running_partition_state(esp_ota_img_states_t state) {
+    esp_ota_mock_has_running_partition = true;
+    esp_ota_mock_running_state = state;
+}
+
 inline const esp_partition_t* esp_ota_get_running_partition() {
-    return nullptr;
+    return esp_ota_mock_has_running_partition ? &esp_ota_mock_running_partition : nullptr;
 }
 inline esp_err_t esp_ota_get_state_partition(const esp_partition_t* /*part*/,
                                              esp_ota_img_states_t* state) {
-    *state = ESP_OTA_IMG_UNDEFINED;
+    *state = esp_ota_mock_running_state;
     return ESP_OK;
 }
 inline const esp_partition_t* esp_ota_get_next_update_partition(
@@ -53,8 +71,13 @@ inline esp_err_t esp_ota_set_boot_partition(
         const esp_partition_t* /*part*/) {
     return ESP_OK;
 }
-inline void esp_ota_mark_app_valid_cancel_rollback() {}
-inline void esp_ota_mark_app_invalid_rollback_and_reboot() {}
+inline void esp_ota_mark_app_valid_cancel_rollback() {
+    esp_ota_mock_mark_valid_called = true;
+    esp_ota_mock_running_state = ESP_OTA_IMG_VALID;
+}
+inline void esp_ota_mark_app_invalid_rollback_and_reboot() {
+    esp_ota_mock_mark_invalid_called = true;
+}
 
 // esp_restart() is in esp_system.h on ESP-IDF but referenced directly here
 inline void esp_restart() {}